+++ narrator.js
@@ -68,7 +68,7 @@ class NarratorMenu extends FormApplication {
  * Primary object used by the Narrator Tools module
  */
 const NarratorTools = {
-    _element: $('<div id="narrator" class="narrator"><div class="narrator-bg"></div><div class="narrator-frame"><div class="narrator-frameBG"></div><div class="narrator-box"><div class="narrator-content"></div></div><div class="narrator-buttons" style="opacity:0;"><button class="NT-btn-pause"></button><button class="NT-btn-close"></button></div></div><div class="narrator-sidebarBG"></div>'),
+    _element: $('<div id="narrator" class="narrator"><div class="narrator-bg"></div><div class="narrator-frame"><div class="narrator-frameBG"></div><div class="narrator-box"><div class="narrator-content"></div></div><div class="narrator-buttons" style="opacity:0;"><button class="NT-btn-pause"></button><button class="NT-btn-close"></button></button><button class="NT-btn-clipboard"></button></div></div><div class="narrator-sidebarBG"></div>'),
     /**
      * Here is where a custom speaker is stored, if you change the value to anything other than '' the next messages will speak as such
      */
@@ -133,15 +133,11 @@ const NarratorTools = {
                     btn[0].classList.remove('active');
             }
         }
-        /**Then, we control the narration's behavior */
-        const is_narrator = game.user.hasPermission('SETTINGS_MODIFY') && game.user.role >= game.settings.get('narrator-tools', 'PERMNarrate');
         /**If a narration had ocurred and the display now is still on, turn it off */
         if (!narration.display && this.elements.content[0].style.opacity === '1') {
             this.elements.BG.height(0);
-            if (is_narrator) {
-                this.elements.buttons[0].style.opacity = '0';
-                this.elements.buttons[0].style.visibility = 'hidden';
-            }
+            this.elements.buttons[0].style.opacity = '0';
+            this.elements.buttons[0].style.visibility = 'hidden';
         }
         /**If the message suddenly disappears, turn off the opacity */
         if (!narration.message) {
@@ -169,7 +165,7 @@ const NarratorTools = {
                             duration = scroll_duration + 4500 * duration_multiplier;
                         }
                     }
-                    if (is_narrator) {
+                    if (this.isNarrator) {
                         if (this._timeouts.narrationCloses) {
                             clearTimeout(this._timeouts.narrationCloses);
                             this._timeouts.narrationCloses = 0;
@@ -190,12 +186,10 @@ const NarratorTools = {
                     this.elements.content[0].style.top = '0px';
                     const height = Math.min(this.elements.content.height() ?? 0, 310);
                     this.elements.BG.height(height * 3);
-                    if (is_narrator) {
-                        this.elements.buttons[0].style.opacity = '1';
-                        this.elements.buttons[0].style.visibility = 'visible';
-                        this.elements.buttons[0].style.top = `calc(50% + ${60 + height / 2}px)`;
-                        this._updateStopButton(game.settings.get('narrator-tools', 'NarrationStartPaused'));
-                    }
+                    this.elements.buttons[0].style.opacity = '1';
+                    this.elements.buttons[0].style.visibility = 'visible';
+                    this.elements.buttons[0].style.top = `calc(50% + ${60 + height / 2}px)`;
+                    this._updateStopButton(game.settings.get('narrator-tools', 'NarrationStartPaused'));
                     this._timeouts.narrationOpens = 0;
                     Hooks.call('narration', narration);
                 }, 500);
@@ -317,11 +311,14 @@ const NarratorTools = {
             buttons: this._element.find('.narrator-buttons'),
             buttonPause: this._element.find('.NT-btn-pause'),
             buttonClose: this._element.find('.NT-btn-close'),
+            buttonCopy: this._element.find('.NT-btn-clipboard'),
         };
         this._updateBGColor();
         this._updateBGImage();
         this._fitSidebar();
         $('body').append(this._element);
+        // Check if the user can Narrate
+        this.isNarrator = game.user.hasPermission('SETTINGS_MODIFY') && game.user.role >= game.settings.get('narrator-tools', 'PERMNarrate');
         // @ts-ignore
         this._menu = new ContextMenuNT({
             theme: 'default',
@@ -346,8 +343,8 @@ const NarratorTools = {
                 },
             ],
         });
-        $(document.getElementById('chat-log')).on('click', '.message.narrator-chat', NarratorTools._onClickMessage.bind(NarratorTools));
-        NarratorTools.elements.buttonPause.on('click', () => {
+        $(document.getElementById('chat-log')).on('click', '.message.narrator-chat', this._onClickMessage.bind(NarratorTools));
+        this.elements.buttonPause.on('click', () => {
             const pause = !NarratorTools.sharedState.narration.paused;
             NarratorTools.sharedState.narration = {
                 ...NarratorTools.sharedState.narration,
@@ -355,12 +352,21 @@ const NarratorTools = {
             };
             NarratorTools._updateStopButton(pause);
         });
-        NarratorTools.elements.buttonClose.html(`<i class="fas fa-times-circle"></i> ${game.i18n.localize('Close')}`);
-        NarratorTools.elements.buttonClose.on('click', NarratorTools._narrationClose);
-        NarratorTools._loadFont(game.settings.get('narrator-tools', 'WebFont'));
-        NarratorTools._updateContentStyle();
+        this.elements.buttonClose.html(`<i class="fas fa-times-circle"></i> ${game.i18n.localize('Close')}`);
+        this.elements.buttonClose.on('click', this._narrationClose);
+        this.elements.buttonCopy.html(`<i class="fas fa-clipboard"></i> ${game.i18n.localize('NT.Copy')}`);
+        this.elements.buttonCopy.on('click', () => {
+            navigator.clipboard.writeText(this.elements.content[0].innerText);
+            ui.notifications.info(game.i18n.localize('NT.CopyClipboard'));
+        });
+        if (!this.isNarrator) {
+            this.elements.buttonPause[0].style.display = 'none';
+            this.elements.buttonClose[0].style.display = 'none';
+        }
+        this._loadFont(game.settings.get('narrator-tools', 'WebFont'));
+        this._updateContentStyle();
         this._controller(game.settings.get('narrator-tools', 'sharedState'));
-        NarratorTools._pause();
+        this._pause();
         document.addEventListener('contextmenu', (ev) => {
             if (ev.target.classList.contains('editor-content') || $(ev.target).parents('div.editor-content').length) {
                 const time = this._menu.isOpen() ? 100 : 0;
@@ -695,6 +701,12 @@ const NarratorTools = {
         };
         /**If the message is a narration, start the protocol */
         if (type == 'narration') {
+            const messageStripped = message
+                .replaceAll('\n', '')
+                .replace(/<(?:\/p|br)[^>]*>/g, '\n')
+                .replace(/<[^>]+>/g, '')
+                .replaceAll('\n', '<br>')
+                .replace(/<br>$/g, '');
             const narration = new Promise((resolve) => {
                 Hooks.once('narration_closes', (narration) => {
                     resolve(narration.message == message);
@@ -711,7 +723,7 @@ const NarratorTools = {
             let state = {
                 id: this.sharedState.narration.id + 1,
                 display: true,
-                message: message,
+                message: messageStripped,
                 paused: game.settings.get('narrator-tools', 'NarrationStartPaused'),
             };
             this.sharedState.narration = state;
@@ -722,6 +734,7 @@ const NarratorTools = {
     },
     /**Shortcuts for easy access of the elements of the module */
     elements: {},
+    isNarrator: false,
     /**
      * Returns the calculated duration a string of length size would have
      * @param length    The lenght of the string